split up lexical e2e
